Scaffold Specs 1–5 harness with clean APIs and TODO stubs.
Add book/ingest/marketdata/replay/bench package surfaces and Spec 1–5 docs so matching and concurrency work can be filled in behind labeled TODOs without changing Spec 6 behavior.

+++ internal/strategy/fakefeed_test.go
@@ -19,7 +19,7 @@ func Trade(symbol string, price, qty int) TradeEvent {
 }
 
 func OwnFillTrade(symbol string, price, qty int, resting OrderID) TradeEvent {
-	id := resting
+	id := uint64(resting)
 	return TradeEvent{
 		Symbol: symbol, Price: price, Quantity: qty,
 		RestingOrderID: &id,
